Fix Stdlib.Platform.Strings on Wasm backend

+++ backend-wasm/runtime/src/shed_platform.h
@@ -13,6 +13,6 @@ typedef uint32_t ShedSize;
 
 typedef uint32_t ShedTagValue;
 
-typedef uint64_t ShedUnicodeScalar;
+typedef uint32_t ShedUnicodeScalar;
 
 #endif
